Moved constant from Tools to AddonsDataProvider
